Issue status check commands after write mem region commands
This is less important for write mem region, but will be very important for read mem region and other commands that need to allow the VB sufficient time to execute. The protocol handling code on the VB isn't the nicest since it involves some loops that don't time out, but I'm not sure what the best approach here is to get around that, and ideally, it's only a problem if the protocol is actually violated and/or something goes wrong, and currently the whole rest of the stack is robust enough to handle and report this, so I think it's ok for now. We'll fix it if it becomes an issue later.